requirements.txt
python-dotenv==1.0.0
PyAutoGUI=0.9.54
opencv-python==4.8.1.78